DRF/DRF-Backend/requirements.txt
asgiref==3.2.10
coverage==5.3
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
pytz==2020.1
sqlparse==0.3.1
